Base functionalities for Workout are done

+++ GymLogger.Api/GymLogger.Api.Client/Pages/Workouts/Index.razor.cs
@@ -1,8 +1,10 @@
 ﻿using GymLogger.Api.Client.Components;
+using GymLogger.Api.Client.Components.Dialog;
 using GymLogger.Shared.Models.Paging;
 using GymLogger.Shared.Models.Workout;
 using GymLogger.Shared.Services;
 using Microsoft.AspNetCore.Components;
+using Microsoft.FluentUI.AspNetCore.Components;
 
 namespace GymLogger.Api.Client.Pages.Workouts;
 
@@ -11,7 +13,7 @@ public partial class Index : BaseComponent
     [Inject] public required IWorkoutApiService WorkoutApiService { get; set; }
 
     private PagedResponseDto<WorkoutDto> _pagedResponseDto = new();
-    private WorkoutPagedRequestDto _pagedRequestDto = new() { SortColumn = "Date" };
+    private WorkoutPagedRequestDto _pagedRequestDto = new() { SortColumn = "Date", SortDirection = Common.Enums.SortDirection.Descending };
 
     private bool sortMenuOpen = false;
     private bool filterOpen = false;
@@ -36,16 +38,35 @@ public partial class Index : BaseComponent
 
     private async Task DeleteAsync(WorkoutDto dto)
     {
-        try
+        var dialog = await DialogService.ShowDialogAsync<DeleteDialog>("Želite li stvarno obrisati ovaj trening?", new DialogParameters()
         {
-            await WorkoutApiService.DeleteAsync(dto.Id);
-            await this.LoadDataAsync();
-            base.ToastService.ShowSuccess("Trening uspješno obrisan.");
-        }
-        catch (Exception)
+            Title = $"Brisanje treninga",
+            PreventDismissOnOverlayClick = true,
+            PreventScroll = true,
+        });
+
+        var result = await dialog.Result;
+        if (!result.Cancelled)
         {
-            base.ToastService.ShowError("Trening nije obrisan");
-            throw;
+            try
+            {
+                await WorkoutApiService.DeleteAsync(dto.Id);
+                await this.LoadDataAsync();
+                base.ToastService.ShowSuccess("Trening uspješno obrisan.");
+            }
+            catch (Exception)
+            {
+                base.ToastService.ShowError("Trening nije obrisan");
+                throw;
+            }
         }
+    }
+
+    private async Task OnSortMenuItemClicked(string sortColumn, Common.Enums.SortDirection sortDirection)
+    {
+        this._pagedRequestDto.SortColumn = sortColumn;
+        this._pagedRequestDto.SortDirection = sortDirection;
+        this.sortMenuOpen = false;
+        await this.LoadDataAsync();
     }
 }
